requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.23
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
slack-sdk>=3.26.1
pydantic>=2.5.2
aiohttp>=3.9.1
python-dateutil>=2.8.2
tenacity>=8.2.3
loguru>=0.7.2
google-auth-httplib2>=0.1.1
google-api-python-client-stubs>=1.18.0
aiosqlite>=0.19.0
greenlet>=3.0.1
pydantic-settings>=2.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
